fix mimxrt10xx, rename k32l2 to kinetis_k32l2
