+++ ports/lpc55/Makefile
@@ -1,5 +1,5 @@
 # List of git submodules that is included as part of the UF2 version
-GIT_SUBMODULES = nxp/mcux-sdk sct_neopixel tinyusb
+GIT_SUBMODULES = tinyusb
 
 include ../make.mk
 include port.mk
